+++ app/src/main/java/com/antonio/samir/meteoritelandingsspots/common/ui/extension/StringExtension.kt
@@ -1,7 +1,7 @@
 package com.antonio.samir.meteoritelandingsspots.common.ui.extension
 
 import java.text.NumberFormat
-import java.util.*
+import java.util.Locale
 
 fun String?.convertToNumberFormat(defaultValue: String): String {
     return if (this.isNullOrBlank()) {
