+++ app/src/main/java/com/antonio/samir/meteoritelandingsspots/data/local/database/MeteoriteDao.kt
@@ -1,7 +1,11 @@
 package com.antonio.samir.meteoritelandingsspots.data.local.database
 
 import androidx.paging.PagingSource
-import androidx.room.*
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import androidx.room.Update
 import com.antonio.samir.meteoritelandingsspots.data.repository.model.Meteorite
 import kotlinx.coroutines.flow.Flow
