+++ designsystem/src/main/java/com/antonio/samir/meteoritelandingsspots/designsystem/ui/components/AddressProgress.kt
@@ -3,13 +3,19 @@ package com.antonio.samir.meteoritelandingsspots.designsystem.ui.components
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ColorFilter
@@ -81,5 +87,81 @@ fun AddressProgresssLightPreview() {
 fun AddressProgresssDarkPreview() {
     MeteoriteLandingsTheme(darkTheme = true) {
         AddressProgress(10.0f, Modifier)
+    }
+}
+
+
+@Preview("Progress Light")
+@Composable
+fun ProgressLightPreview() {
+
+    MeteoriteLandingsTheme(darkTheme = false) {
+
+        var state by remember { mutableStateOf(0.0f) }
+
+        Column(
+            modifier = Modifier
+                .padding(16.dp)
+                .fillMaxWidth()
+        ) {
+            AddressProgress(
+                progress = state,
+                modifier = Modifier
+            )
+            Button(
+                onClick = { state = 10.0f },
+                modifier = Modifier
+                    .align(Alignment.CenterHorizontally)
+                    .padding(8.dp)
+            ) {
+                Text(text = "Show")
+            }
+            Button(
+                onClick = { state = 0.0f },
+                modifier = Modifier
+                    .align(Alignment.CenterHorizontally)
+                    .padding(8.dp)
+            ) {
+                Text(text = "Hide")
+            }
+        }
+    }
+}
+
+
+@Preview("Progress Dark")
+@Composable
+fun ProgressLightDark() {
+
+    MeteoriteLandingsTheme(darkTheme = true) {
+
+        var state by remember { mutableStateOf(0.0f) }
+
+        Column(
+            modifier = Modifier
+                .padding(16.dp)
+                .fillMaxWidth()
+        ) {
+            AddressProgress(
+                progress = state,
+                modifier = Modifier
+            )
+            Button(
+                onClick = { state = 10.0f },
+                modifier = Modifier
+                    .align(Alignment.CenterHorizontally)
+                    .padding(8.dp)
+            ) {
+                Text(text = "Show")
+            }
+            Button(
+                onClick = { state = 0.0f },
+                modifier = Modifier
+                    .align(Alignment.CenterHorizontally)
+                    .padding(8.dp)
+            ) {
+                Text(text = "Hide")
+            }
+        }
     }
 }
