+++ designsystem/src/main/java/com/antonio/samir/meteoritelandingsspots/designsystem/ui/components/Shimmer.kt
@@ -1,0 +1,38 @@
+package com.antonio.samir.meteoritelandingsspots.designsystem.ui.components
+
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.valentinilk.shimmer.ShimmerBounds
+import com.valentinilk.shimmer.rememberShimmer
+import com.valentinilk.shimmer.shimmer
+
+
+@Composable
+fun Shimmer(modifier: Modifier = Modifier) {
+    val shimmer = rememberShimmer(shimmerBounds = ShimmerBounds.Window)
+    Box(
+        modifier = modifier
+            .padding(12.dp)
+            .width(152.dp)
+            .height(20.dp)
+            .shimmer(customShimmer = shimmer),
+        contentAlignment = Alignment.Center
+    ) {
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(24.dp)
+                .background(Color.LightGray)
+        )
+    }
+}
